+++ run.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+# termatplotlib launcher for Unix/macOS
+# Usage: ./run.sh bar --labels "A B C" --data "10 20 30"
+
+python -m termatplotlib "$@"